Improve visibility in build process
`make build` now prints the nginx version embedded in the image when it
is done building.

Additionally, there is a `make latest` that does a docker pull of the
nginx:latest image so that the dproxy image is based on latest and
greatest upstream.

Usage is now either:

  make latest build

or

  make latest build push

(unless you know for a fact that you do not want to update the local
 nginx image to the most recent `latest` tag...)